youtube_chatbot_backend/requirements.txt
# LangChain Core
langchain
langchain-core

# OpenAI Integration
langchain-openai
openai

# Environment Variable Management
python-dotenv